[PLAT-15225] Verify there is no running master on nodes selected for master replacement
Summary:
This adds remote call to check process status for master. As this is run in findReplacementMaster which is invoked in a transaction block, call site is changed to precheck to not hold up the transaction. This changes the object references. So, node name is returned
to avoid making mistakes by using the node object to update universe record. The call site move means master state also needs to be checked because isMaster can be false.

Test Plan:
1. Created a 5 node RF 3 cluster.
2. Stopped node with master spins up another master on node.
3. Started the node and started master on the tserver only node.
4. Stopped this new master node. Precheck failed.
5. Stoppped the master from the tserver only node.
6. Stopped the new master node. This time, it succeeded.
7. Repeat the same for remove node.
8. Tested master failover by stopping a master VM.

{F287918}

Reviewers: cwang, sanketh, yshchetinin

Reviewed By: cwang, yshchetinin

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38254