[#22523] YSQL, ASH: Instrument all RPCs from PgClient to local TServer
Summary:
This diff instruments all the RPCs defined in pg_client.proto
except the following -
- GetCatalogMasterVersion - Deprecated
- GetReplicationSlotStatus - Deprecated
- SetActiveSubTransaction - Deprecated
- CreateSequencesDataTable - Deprecated in PG
- GetDatabaseInfo - This RPC is fired while creating a PG session, before ASH metadata is set, so it won't be sampled
- IsInitDbDone - This RPC is fired with bootstrap mode when ASH collector is not yet up, again, cannot be sampled
- ActiveSessionHistory - This RPC is only fired from the ASH collector which is a background worker and background workers are not yet sampled (#23901)
- CronSetLastMinute - This is also fired from a background worker
- CronGetLastMinute - This is also fired from a background worker
- FetchData - This is an async RPC and we don't seem to wait for it, the callback is called upon it's completion
- Heartbeat - This is also an async RPC and we don't wait for it

This diff does only covers instrumentation at YSQL layer, and not
at the TServer layer. All the applicable sync RPCs are instrumented
with this wait event - WaitingOnTServer, and the RPC name is stored
as an int, and we fetch the RPC name from the enum `PggateRPC`
when ASH is queried. The ASH metadata is also sent along with the
RPC to TServer if ASH is enabled.

A new variable `yb_rpc_code` is added to PGPROC struct to store the
pggate rpc enum. While sampling, we store this in the `aux_info` field
in `YBCAshSample`. `aux_info` is of 16 bytes, we store the pggate
rpc enum in the first byte, and the second byte is set to zero. We use
this rule to distinguish the aux info for PG and DocDB samples, and
fetch the rpc name from the `PggateRPC` enum

**Upgrade/Rollback safety:** Safe to upgrade/rollback, if ASH metadata
is not found in the PB, then metadata in the threadlocal wait state object
is not updated.

If ASH metadata is sent, and the other tserver is still on old version, it will
simply ignore and continue doing it's work
Jira: DB-11450

Test Plan:
Updated old test
./yb_build.sh --cxx-test wait_states-itest --gtest_filter WaitStateITest/AshTestVerifyPgOccurrence.VerifyWaitStateEntered/kWaitingOnTServer

New unit tests in pg_ash-test

The test testPgAuxInfo is removed as we now have wait event aux for
PG samples too.

Reviewers: dmitry

Reviewed By: dmitry

Subscribers: yql, amitanand, hbhanawat, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38309