[#24040] YSQL: Simplify the PatchStatus function
Summary:
The is no reason to call the `PgSession::LoadTable` method in the `PatchStatus` function to get proper relation id.
It is reasonable to sore relation id (`PgTableDesc::pg_table_id`) in the `relations` vector directly instead of `PgTableDesc::relfilenode_id`.

**Note:**
In context of this diff some cosmetic changes in the `BufferableOperations` class are performed:
- rename of some methods to fit codding style
- disallow non-const access to the `operations` and `relations` fields
Jira: DB-12930

Test Plan: Jenkins

Reviewers: fizaa, jhe, myang

Reviewed By: myang

Subscribers: ybase, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38221