Merged PR 1462597: Add apiStrict and apiDeprecationErrors to ignored spec fields
### Does this PR have any customer impact?

### Type (Feature, Refactoring, Bugfix, DevOps, Testing, Perf, etc)

### Does it involve schema level changes? (Table, Column, Index, UDF, etc level changes)

### Are you introducing any new config? If yes, do you have tests with and without them being set?

### Release Note (Refer [Template](../docs/release_notes/pgmongo_template.md))

### Description

We do not support apiVersion, so apiStrict and apiDeprecationErrors can be ignored for now.

----
#### AI description  (iteration 1)
#### PR Classification
Code update to support or ignore specific API parameters.

#### PR Summary
This pull request updates the code to include `apiStrict` and `apiDeprecationErrors` in the list of ignored specification fields, addressing the requirements of the linked work item.
- `commands_common.c`: Added `apiStrict` and `apiDeprecationErrors` to the `IgnoredCommonSpecFields` array.

Related work items: #3463178