[#16487] YSQL: Fix flakey TestPostgresPid test
Summary:
`TestPostgresPid` was flaky, failing ocassionally on Jenkins and failing ~40% of the time on my devserver. The failures came from lines 364 and 380 - waiting for PG to shut down.

The problem is due to Postgres not fully initializing before the kill signal was sent. After restarting the tserver, waiting for Postgres to start allows the cluster to get into a good state.

This change also strengthens the test a bit - not only are the tservers up, but PG is ready to serve clients.
Jira: DB-5890

Test Plan:
Jenkins: test regex: .*TestPostgresPid.*
```
./yb_build.sh --cxx-test pg_wrapper-test --gtest_filter PgWrapperOneNodeClusterTest.TestPostgresPid -n 50
```

Reviewers: kramanathan

Reviewed By: kramanathan

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38225