[#24091] xClusterDDLRepl: Pass automatic_ddl_mode to pollers
Summary:
Storing automatic_ddl_mode in ProducerEntryPB so that it can be passed to xClusterPollers.
This will be used in future diffs to know if we need special handling of ChangeMetadataOps.
Jira: DB-12984

**Upgrade/Rollback safety:**
Adding new field automatic_ddl_mode which is only used going forward and assumed to be false for older releases

Test Plan: Jenkins

Reviewers: hsunder, xCluster, mlillibridge

Reviewed By: hsunder, mlillibridge

Subscribers: mlillibridge, ybase, ycdcxcluster

Differential Revision: https://phorge.dev.yugabyte.com/D38314