[PLAT-15385] Remove ProgressMonitor from Commissioner as it is no longer used
Summary:
ProgressMonitor is unused and unneccessary. This change removes it. Moreover, each subtask group creation does not need to recreate thread pool of the same parameters everytime. This makes sure to create it once and is used until the
end of the parent task to remove the overhead of termination and creation.

Test Plan:
Trivial change:

1. Created universe. 2. Edited the universe. Everything completed fine.

Itests should also verify more.

Reviewers: amalyshev, sanketh, cwang, yshchetinin

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38337