[BACKPORT 2024.1][#23943] YSQL: Fix Bitmap Scan GCC11 crash (follow-up)
Summary:
Original commit: 76a5c975c166f03a4874bc9fdd44181ce7cc52ea / D38318
== Original Summary ==
89b69cfc010fc5c0ef31ce968026d1f42b9a0eae / D38080 attempted to fix the crash in GCC11, but after an initial successful run on Jenkins I made a small change and tested locally, omitting the `--gcc11` flag that was required to reproduce the crash. Since I had already seen a successful Jenkins run and believed the local results were still good, I didn't look closely at the Jenkins results and missed that it had started failing again on gcc11 fastdebug.

The previous hypothesis was that `delete[]`ing the memory caused the `Slice`'s data to point to a garbage value, which could potentially cause errors while erasing the slice from the unordered_set. This proved not to be true, as calling `Slice::Clear` after deleting the memory did not resolve the problem.

The problem is actually mentioned by the [[ https://en.cppreference.com/w/cpp/container/unordered_set | C++ standard ]]
> Container elements may not be modified (even by non const iterators) since modification could change an element's hash and corrupt the container.

Rearranging the instructions so that the modification of the Slice occurs after it is removed from the set solves the problem. This is also the same as the first draft of the previous attempt of a fix, but now the reason is well understood.

Jira: DB-12839

== Differences with the original diff ==
The original diff (D38318) removed the line `slice.Clear();` from the `FreeSlice` function. That line was added by the previous attempt (D38080) to address the crash.
* master and 2024.2 had the `slice.Clear();` line, because 2024.2 was cut after D38080 landed.
* 2024.1 did not have the `slice.Clear()` line because D38080 was not backported to 2024.1.

This diff does not remove the `slice.Clear()` line from `FreeSlice`, because it is not present in this branch.

Test Plan:
Jenkins,

Manually:
```
Build:
./yb_build.sh fastdebug --gcc11
Restart cluster
Run
/*+ Set(yb_enable_bitmapscan true) Set(yb_enable_base_scans_cost_model true) BitmapScan(t) */
EXPLAIN (ANALYZE, SUMMARY OFF, COSTS OFF) SELECT * FROM test_and t WHERE a < 5 AND b < 5;

\watch .1
```

Reviewers: amartsinchyk, smishra

Reviewed By: amartsinchyk, smishra

Subscribers: smishra, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38350