[pg15-upgrade] test: work around performance issue on MacOS
Summary: YB master commit 981415e73ec2b348a16441bfb4cb41f35a8227f5 changed the default value of pg_client_use_shared_memory to true in release builds. This causes a severe performance regression for initdb on MacOS. Modify the test flags to set this flag to false.

Test Plan:
On MacOS 14 arm64 (only known platform with this issue):
  ./yb_build.sh release --sj
  pg15_tests/test_upgrade.sh

Check the master leader's log for the line:
  pg_wrapper.cc:1034] initdb took 23611 ms

The number should be around 24 ms, which it is here. Without this change the number is around 10x that.

Reviewers: fizaa, telgersma

Reviewed By: telgersma

Subscribers: ybase, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38352