[YSQL][conn-mgr] Change the log level in the message during sleep post DDL
Summary: When connection manager is enabled, post DDL we wait for a certain duration for the DDL changes to be notified to other backends. During the wait phase, we are logging some info indicating the waiting time. Changing the log level to `LOG_SERVER_ONLY` from `LOG` as this info need not be propagated to the user ever. This additional log message is also causing the test failure `TestPgRegressTAQO#testPgRegressTAQO`. The change will address both the issues.

Test Plan: Jenkins: test regex: .*TestPgRegressTAQO.*, enable connection manager

Reviewers: devansh.saxena, skumar

Reviewed By: devansh.saxena

Subscribers: yql

Differential Revision: https://phorge.dev.yugabyte.com/D38364