[BACKPORT 2024.2][#21467,#21783, #23667] Docdb: Add sequences support for clone part 1
Summary:
Original commit: 8916c1d1f6bd65f6aa1eb0156110cbe5061e97d9 / D38186
Currently, clone doesn't support sequences in YSQL database. That is, if we try to clone a database which has a sequence, cloning will fail with the following error:
```
ERROR:  Clone operation aborted: Not found (yb/tablet/tablet_metadata.cc:781): Failed to list snapshot: c0d01ed0-0344-49cb-9e69-386b8e75b558: Table <unknown_table_name> (0000ffff0000300080000000000004eb) not found in Raft group 00000000000000000000000000000000
```
The clone was failing in the Generate SnapshotInfo as of time step. More specifically, clone repacks the SnapshotInfo of the suitable snapshot to be consumed in the restore operation.
However, as the snapshot repacked by clone is part of a snapshot schedule, it contains the `sequences_data` table entry i.e: table: `0000ffff0000300080000000000004eb`. The repacking step doesn't expect the `sequences_table` to be inside the snapshot and thus throws the previous error.
Cloning sequences is one of the situations where clone diverge from PITR. In clone, we use Backup/Restore flow to restore sequences while in PITR we use the snapshot of the `sequences_data` table to get the correct value at restore time. The diff fixes the issue by skipping the repacking of the `sequences_data` table and related tablets from the snapshotInfo. The assumption is that we never need to repack the `sequences_data`. In reality, `sequences_data` table only exists in snapshots that are part of a snapshot schedule.
Jira: DB-10350, DB-10673, DB-12578

Test Plan:
./yb_build.sh --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter Colocation/PgCloneTestWithColocatedDBParam.CloneWithSequences/0
./yb_build.sh --cxx-test integration-tests_minicluster-snapshot-test --gtest_filter Colocation/PgCloneTestWithColocatedDBParam.CloneWithSequences/1

Also Enabled the following PITR tests for clone and all of them passed after this diff:
PgsqlSequenceUndoDeletedData
PgsqlSequenceUndoInsertedData
PgsqlSequenceUndoCreateSequence

Example to run one of them:
./yb_build.sh --cxx-test yb-admin-snapshot-schedule-test --gtest_filter ColocationAndRestoreType/YbAdminSnapshotScheduleTestWithYsqlColocationRestoreParam.PgsqlSequenceUndoDeletedData/3
3 and 4 are the parameters for clone (non-colocation and colocation respectively)

Reviewers: asrivastava

Reviewed By: asrivastava

Subscribers: ybase, slingam

Differential Revision: https://phorge.dev.yugabyte.com/D38375