[#21733] YSQL: ParallelAppend and pg_hint_plan
Summary:
The pg_hint_plan unconditionally promoted ParallelAppend in UNION/UNION ALL.
They did not give a good reason, just mentioned that UNION does not have a name, so it is unhintable.
That does not mean it is good to always promote Parallel, in our experience it was bad in some cases.
It does not affect us much while ParallelAppend is disabled by default, but may cause regression in the future.
Hence removing the offending code block.
Jira: DB-10605

Test Plan:
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressParallel#testPgRegressParallel'
./yb_build.sh --java-test 'org.yb.pgsql.TestPgRegressThirdPartyExtensionsPgHintPlan'

Reviewers: mtakahara, tnayak, jason

Reviewed By: mtakahara

Subscribers: jason, yql

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38288