[#24056] CDCSDK: Fix flaky test CDCSDKYsqlTest.TestNonEligibleTableCleanupWithDropTable
Summary:
This test drops an index after a master restart. However, the drop index operation was timing out because it began executing tasks before the tserver registration was complete. As a result, while trying to delete tablets as part of the drop operation, it couldn't find a replica for the index's tablet, causing the operation to repeatedly attempt deleting the tablets and eventually time out.

In failed runs, 250+ occurences of below logs are observed whereas in successful run, maximum of 10 occurences are seen.
```
catalog_manager.cc:6838] Servicing IsDeleteTableDone request for table id 00004000000030008000000000004005: deleting tablets
```

The fix is to wait for the tablet servers to complete their registration with the master after the restart, and then proceed with dropping the index.

Note, this test only fails on 2.20 and not on other branches like master/2024.1. This could be most likely because some changes in mini_cluster/mini_master would have resolved this but it might not have been backported to 2.20.

This test, when added, failed once on 2.20 because we tried to drop the index before master had loaded all the tables after restart. As a result, it failed with SIGSEGV, which is why `WaitUntilCatalogManagerIsLeaderAndReadyForTests()` was added. But, this is no longer required as we now wait for tserver to register which might account for the time taken for master to load all metadata.
Jira: DB-12948

Test Plan: Jenkins: test regex: .*TestNonEligibleTableCleanupWithDropTable.*

Reviewers: sumukh.phalgaonkar, skumar, stiwary

Reviewed By: skumar

Subscribers: ycdcxcluster

Tags: #jenkins-ready

Differential Revision: https://phorge.dev.yugabyte.com/D38382