[PLAT-15432] remove status,sizeInBytes from manifest.json file
Summary:
`Manifest.json` file is created before any of the support bundle components are collected so the size is always 0 and status is always running.
This is the case for all support bundles and so doesn't make sense to include them in the manifest.

Test Plan: created support bundle and verified that the fields are not present in manifest.json file.

Reviewers: #yba-api-review!, skurapati

Reviewed By: skurapati

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38390