[#23730] YSQL: Skip internal savepoint for SET/BEGIN statement in RC isolation
Summary:
After 3494424, the following leads to a crash in debug builds (with the
Tserver gflag yb_enable_read_committed_isolation set to true):

```
BEGIN;
SET TRANSACTION ISOLATION LEVEL SERIALIZABLE;
SELECT 1;
ROLLBACK;

TRAP: FailedAssertion("!(MySerializableXact->pid == 0 || (XactIsoLevel == 3))", File: "../../../../../../../src/postgres/src/backend/storage/lmgr/predicate.c", Line: 3277)
```

This is because of the following reason:

1) The transaction starts in Read Committed isolation since that is the default.
2) The first statement acquires an internal savepoint before execution (as per YB's RC implementation).
3) The isolation of the transaction is changed to SERIALIZABLE.
4) The ROLLBACK results in aborting the sub-transaction followed by the top level transaction. However, aborting the sub-transaction results in executing AtEOXact_GUC which restores the GUC values before the savepoint was set. This revert the isolation level back to read committed. Post this, the abort for the top level transaction eventually ends up calling ReleasePredicateLocks() (predicate locks are taken only in serializable isolation level of vanilla PG, but YB takes them too since those code paths haven't been disabled). Since the isolation level GUC has reverted back to read committed, the following assertion fails:

```
Assert(MySerializableXact->pid == 0 || IsolationIsSerializable());
```

To fix this, we will skip registering an internal savepoint in Read
Committed isolation for SET statements. This is also done for BEGIN
statements with code comments to explain the reasoning.

Given this change, we no longer need the logic in 3494424. So, this diff
removes the usage of YbHasOnlyInternalRcSubTransactions().
Jira: DB-12634

Test Plan:
./yb_build.sh --java-test org.yb.pgsql.TestPgTransactions#testSetTransactionIsolationInRC
./yb_build.sh --java-test org.yb.pgsql.TestPgRegressWaitQueues#testPgRegress

Tests that were added in commit 79888cf1 which was built on top of 3494424f (changes of which have been removed):
./yb_build.sh --cxx-test 'TEST_F(PgMiniTest, FollowerReads) {'
./yb_build.sh --java-test org.yb.pgsql.TestPgFollowerReads#testSetIsolationLevelsWithReadFromFollowersSessionVariable

Reviewers: mtakahara, jason, patnaik.balivada

Reviewed By: mtakahara, patnaik.balivada

Subscribers: patnaik.balivada, amitanand, yql

Differential Revision: https://phorge.dev.yugabyte.com/D38396