[pg15-upgrade] test: create partition table test
Summary:
Create three types of partitioned tables:
1. Range partitioned
2. Hash partitioned
3. List partitioned

Validate that their output is correct at each stage of the upgrade process. From a PG15 node, insert a new row. From a PG11 node, check that the row is present and delete it.

When the upgrade completes, try creating new partition tables and inserting new values. Validate that the new values appear in selects to both the partitioned table and the correct partition table.

Test Plan:
```
for i in {1..10}
    do pg15_tests/test_upgrade_partition.sh &> /dev/null
    echo $?
done
```

Reviewers: tfoucher, fizaa

Reviewed By: tfoucher

Differential Revision: https://phorge.dev.yugabyte.com/D38373