[PLAT-14954] added support for systemd-timesyncd
Summary:
both clock-sync.sh and health checks/metrics for clock drift now support
using systemd-timesyncd. For drift metrics, we can only return if the
system is in-sync with the remote, as timesycd is only SNTP, not full NTP,
and only does time-steps to update the clock when a drift is detected.

Test Plan: tested timesyncd works

Reviewers: muthu, yash.priyam, amalyshev

Reviewed By: amalyshev

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38440