[pg15-upgrade] test: enable upgrade tests for pg15
Summary:
This change adds support for upgrade across ysql major versions in the upgrade tests.
This is currently limited to upgrades from `2.20.2.4`.

Adds `ysql_major_version` to `VersionInfoPB`. The JSON file already has this info, and this change merely adds it to the PB which is returned in `GetStatus` RPC.

Using the `ysql_major_version` info from the old and new versions the test determines `is_ysql_major_version_upgrade_ `.
If `is_ysql_major_version_upgrade_ ` is enabled we now perform the following extra steps:
- Set `yb_enable_expression_pushdown` before restarting processes in new version and clear if after all processes are on the same version.
- Set `ysql_upgrade_postgres_port` on yb-master with unique pg ports.
- Set `TEST_online_pg11_to_pg15_upgrade` on yb-master before invoking ysql major upgrade API, and clear it on Finalize/Rollback.
- Run `PerformYsqlMajorVersionUpgrade`, `FinalizeYsqlMajorVersionUpgrade` and `RollbackYsqlMajorVersion` functions.

The `GetClosestLiveTserver` function has been added to `CatalogManager` in the master branch. That change has not reached `pg15-upgrade` yet, so manually copying this function over. This function is more stable and has tests, so removing `YsqlInitDBAndMajorUpgradeHandler::GetClosestLiveTserverAddress`.

Test Plan: Jenkins

Reviewers: tfoucher

Reviewed By: tfoucher

Subscribers: telgersma, ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38435