[PLAT-15493] Failover fails post switchover in a replication setup
Summary: With recent changes, YBA creates the PITR config on both universes during the create DR config task and the switchover task should not delete that pitr config.

Test Plan: Created a DR config between two universes and run switchover multiple times and made sure that the pitr is not deleted and a failover task can be run after switchover.

Reviewers: cwang, vbansal, sanketh

Reviewed By: cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38482