[#24164] xClusterDDLRepl: Skip CreateSequencesDataTable for AlterUniverseReplication
Summary:
Currently we attempt to CreateSequencesDataTable for all SetupUniverseReplication requests, which also includes AlterUniverseReplication requests to add in new tables. This causes unnecessary work as we have to go through a number of checks in CreateTable before realizing there's a duplicate table and returning.

Skipping this check for all alter replication requests as the sequences_data table is a global table, so would have been created during the original setup.
Jira: DB-13051

Test Plan: Jenkins

Reviewers: mlillibridge, hsunder, xCluster

Reviewed By: mlillibridge, hsunder

Subscribers: ybase

Differential Revision: https://phorge.dev.yugabyte.com/D38484