[BACKPORT 2.20][PLAT-15496][PLAT-15385] Handle throwable in TaskExecutor
Summary:
These two are related changes. The latter depends on the first change.

Original diffs -
1. https://phorge.dev.yugabyte.com/D38337 (82b2dd695dc8e4d76270bdaf4557c124d50482e5)
2. https://phorge.dev.yugabyte.com/D38490 (4f53e36023163af0a17ab5c4da16e04ad7f4cf14)

[PLAT-15385] Remove ProgressMonitor from Commissioner as it is no longer used

ProgressMonitor is unused and unneccessary. This change removes it. Moreover, each subtask group creation does not need to recreate thread pool of the same parameters everytime. This makes sure to create it once and is used until the
end of the parent task to remove the overhead of termination and creation.

[PLAT-15496] Handle throwable in TaskExecutor

Catch and update task but propagate Error as it is to report serious errors.

Test Plan:
Trivial change:

1. Created universe. 2. Edited the universe. Everything completed fine.

Itests should also verify more.

Trivial change. UT passed. Itests should pass.

Reviewers: amalyshev, sanketh, cwang, yshchetinin, vkumar, anijhawan

Reviewed By: cwang

Subscribers: svc_phabricator, yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38643