[PLAT-15519]: Update xCluster sync to remove tables from YBA DB
Summary: Updated xCluster sync to remove tables from YBA DB if table is removed from replication group and do not check it presence on source universe tables.

Test Plan:
Tested manually by dropping the table on target universe and verified that tables is listed as Extra table on source if it still exists on source universe.
Tested on both DR and Non-DR replications.

Reviewers: hzare, cwang, sanketh

Reviewed By: hzare, cwang

Subscribers: yugaware

Differential Revision: https://phorge.dev.yugabyte.com/D38529